InputForm自律
