+++ src/javascripts/typings.d.ts
@@ -12,6 +12,12 @@ declare global {
     };
   }
 
+  declare module 'react-redux' {
+    interface DefaultRootState extends State {}
+    function useDispatch<TDispatch = Dispatch<Action>>(): TDispatch;
+    function useStore<S = DefaultRootState>(): Store<S, Action>;
+  }
+
   interface Window {
     // https://github.com/zalmoxisus/redux-devtools-extension
     __REDUX_DEVTOOLS_EXTENSION__?(): StoreEnhancer;
